clean: Refactor logic
