+++ src/test/scala/com/codacy/clangtidy/ClangTidyReportParserSpecs.scala
@@ -4,14 +4,14 @@ import java.nio.file.Paths
 
 import org.scalatest.{Matchers, WordSpec}
 
-class ReportParserSpecs extends WordSpec with Matchers {
+class ClangTidyReportParserSpecs extends WordSpec with Matchers {
 
   "ReportParser::parse" should {
     "parse simple result" in {
       val line =
         "/src/main.c:10:6: error: expected expression [readability-else-after-return]"
 
-      new ReportParser().parse(Seq(line)) should be(
+      new ClangTidyReportParser().parse(Seq(line)) should be(
         Seq(
           ClangTidyResult(
             Paths.get("/src/main.c"),
@@ -29,7 +29,7 @@ class ReportParserSpecs extends WordSpec with Matchers {
       val line =
         "/src/main.c:15:3: error: do not use 'else' after 'return' [readability-else-after-return,-warnings-as-errors]"
 
-      new ReportParser().parse(Seq(line)) should be(
+      new ClangTidyReportParser().parse(Seq(line)) should be(
         List(
           ClangTidyResult(
             Paths.get("/src/main.c"),
@@ -68,7 +68,7 @@ class ReportParserSpecs extends WordSpec with Matchers {
         )
       )
 
-      new ReportParser().parse(line) should be(expected)
+      new ClangTidyReportParser().parse(line) should be(expected)
     }
 
     "ignore lines that don't match " in {
@@ -78,7 +78,7 @@ class ReportParserSpecs extends WordSpec with Matchers {
         "Suppressed 17 warnings (17 in non-user code)."
       )
 
-      new ReportParser().parse(line) should be(empty)
+      new ClangTidyReportParser().parse(line) should be(empty)
     }
 
     "parse a full output correctly" in {
@@ -125,7 +125,7 @@ class ReportParserSpecs extends WordSpec with Matchers {
         )
       )
 
-      new ReportParser().parse(output.split("\n").toSeq) should be(expected)
+      new ClangTidyReportParser().parse(output.split("\n").toSeq) should be(expected)
     }
   }
